Refactored options to objects with keys
Numeric indexes with disabled options were started to be messy. E.g.
option 4 is disabled, and next option 5 doesn't exist, as it's the
"new" 4. Changed this by switching over to option names as keys.

It also makes the template a bit more readable, as values are adressed
with their keys and not indexes. ( opt[2] -> opt.refresh )

Compatible with existing cookies.